+++ Builds/VisualStudio2013/RippleD.vcxproj
@@ -2232,8 +2232,6 @@
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\crypto\ECDSACanonical.h">
     </ClInclude>
-    <ClInclude Include="..\..\src\ripple\crypto\ECDSAKey.h">
-    </ClInclude>
     <ClInclude Include="..\..\src\ripple\crypto\ECIES.h">
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\crypto\ec_key.h">
@@ -2261,6 +2259,8 @@
     <ClCompile Include="..\..\src\ripple\crypto\impl\ECDSAKey.cpp">
       <ExcludedFromBuild>True</ExcludedFromBuild>
     </ClCompile>
+    <ClInclude Include="..\..\src\ripple\crypto\impl\ECDSAKey.h">
+    </ClInclude>
     <ClCompile Include="..\..\src\ripple\crypto\impl\ECIES.cpp">
       <ExcludedFromBuild>True</ExcludedFromBuild>
     </ClCompile>
